requirements.txt
pyyaml
python-dotenv
ultralytics
opencv-python
torch
torchvision